# docs/swagger.yaml
basePath: /
definitions:
...
  server.loginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
...
  server.signupRequest:
    properties:
      city:
        example: Mumbai
        type: string
      country:
        example: India
        type: string
      email:
        example: user@example.com
        type: string
      mobile_number:
        example: +91-9876543210
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        type: string
      state:
        example: Maharashtra
        type: string
    required:
    - email
    - name
    - password
    type: object
  server.simpleResponse:
    properties:
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  server.verifyEmailRequest:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: user@example.com
        type: string
    required:
    - code
    - email
    type: object
...
host: localhost:5001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Type "Bearer" followed by a space and JWT token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Front Insight API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.simpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.simpleResponse'
      summary: User login
      tags:
      - Authentication
...
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with email verification
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.signupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.simpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.simpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.simpleResponse'
      summary: Register a new user
      tags:
      - Authentication
...
  /verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with the verification code sent during signup
      parameters:
      - description: Email verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.verifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.simpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.simpleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.simpleResponse'
      summary: Verify user email
      tags:
      - Authentication
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"